+++ http/gorilla/module.go
@@ -1,0 +1,9 @@
+package gorilla
+
+import (
+	"github.com/gorilla/mux"
+	"go.uber.org/fx"
+)
+
+// Module is an fx compatible module.
+var Module = fx.Provide(mux.NewRouter, NewHandler)